# server/services/status/index.js
"use strict";
const express = require("express");
const { connectToMongoDB } = require("../../dbconfig");
const router = express.Router();

router.post("/create_status", async (req, res) => {
  let db;
  try {
    // Connect to MongoDB
    db = await connectToMongoDB();

    const { statusName } = req.body;

    // Validate required fields
    if (!statusName) {
      return res
        .status(400)
        .json({ success: false, error: "All fields are required" });
    }

    // Create a new status object
    const newStatus = {
      statusName,
      createdAt: new Date(), // Use the current date and time
    };

    // Insert the new status into the "status" collection
    await db.collection("status").insertOne(newStatus);

    res
      .status(201)
      .json({ success: true, message: "Status added successfully" });
  } catch (ex) {
    console.error("Database query error:", ex); // Log the error for debugging
    res.status(500).json({ success: false, error: ex.message });
  }
  // No need to close the connection here; it should remain open for future requests
});

router.get("/status", async (req, res) => {
  let db;
  try {
    // Connect to MongoDB
    db = await connectToMongoDB();

    // Query the "status" collection
    const status = await db.collection("status").find({}).toArray();

    res.status(200).json({ success: true, data: status });
  } catch (ex) {
    console.error("Database query error:", ex); // Log the error for debugging
    res.status(500).json({ success: false, error: ex.message });
  }
  // No need to close the connection here; it should remain open for future requests
});

module.exports = router;